Experimental sound resource caching stuff
The real reason for this is not having to write lots of SND files in the game file location
for Hugo games.

A more Glk-compliant way to do it would have been to create a new blorb in memory,
but this is a lot simpler, basically just an array of structs with pointers to resource data.
It wastes some memory but hopefully causes no major new leaks.