Add style to the protocol when drawing images
If images don't have the same paragraph style as the rest of the paragraph,
all paragraph settings are disregarded by the text system. So we must apply the
paragraph style to the image, but we have abstracted away the concept of window
style in the Cocoa window manager. We now include the style in the draw image
command. This requires a new struct to hold the extra parameter.