Preliminary autoscroll
Scroll down one screen if a command takes a long time to complete
(more than half a second.)
The purpose of this is to scroll down on testing commands such as TEST ME in
Counterfeit Monkey. But it seem to trigger elsewhere as well for, such as during
Kerkerkruip animations, which isn't ideal.