Decrease delay when entering fullscreen
This happens when a game is closed in fullscreen and then opened:
It is opened in windowed mode, then enter fullscreen after a delay that is supposed
to be long enough to allow everything to be in its right (windowed) place.

Hopefully this decreased delay does not cause trouble on slower machines.