Preferences: Make deleting themes messier
Basically we try harder to replace removed themes with
the theme they were based on. So if a game uses theme
"Automap (Modified)" and "Automap (Modified)" is deleted,
the game will be moved to use theme "Automap".

We have the relations defaultParent and defaultChild that
takes care of this in theory, but we try to guess from the
file name first, because many people will have libraries
created by older versions.