Don't bring game window to front if showing file access dialog
When a game is opened by double clicking the game file, Spatterlight will move its
window to the front after a delay, to make sure it does not end up behind the library window.

When the user has not granted access to the game folder, a modal file dialog will show, and
we don't want the game window to be pushed in front of this dialog. We avoid this with a flag.