TableViewController: Redo edit on click and wait a bit
A little less code, a little less willing to edit. However, this
introduces a slight inconsistency that I don't know how to fix:
a row has to be selected first for edit-on-click-and-wait to work,
but if you are editing a field, clicking outside it on another field in the
sam row will activate edit on wait in the new field right away, which
feels wrong.