ZMenu: Bump up the maximum length of strings to be checked
The the text string of a grid window will grow rapidly if the window
is enlarged, and you quickly get 5000 characters and above with standard
fonts in a screen-wide window. It is still easy to get above 10 000 characters
too by making the font smaller, but we have to draw a line somewhere.

Also bumping up the limit of the Bureaucracy form check.