Adjust line height for graphic font
In Journey the bottom would get cut off with certain fonts
if machine was set to Amiga, which draws a border using
the graphic font. In Shogun the two-line status bar would
look squeezed and cut off with the CGA and VGA fonts in
the DOSBox and MS-DOS themes.

This seems to fix that, at least on Sequoia.