GlkController: More sanity checks for bad window sizes
There is less need for this now, but we should guard against
autosaves made with old buggy versions. We also want to be able
to resize a window to minimum allowed size, autosave and have it
restore to the minimum size. Thus we have to adjust the minimum
constants slightly.