InfoController: Kliudge to avoid system game mode notification
Whenever we create a fullscreen window, the system will issue a notification
about entering game mode. That is kind of confusing when we just use the window
for a temporary animation, so we get around it by making the window one point smaller
than fullscreen. Presumably this means that we previously entered a high-performance  game mode,
and we now no longer do, which might make the animation less smooth. Probably still worth it.