Don't repeat old text if there is no new
The GlkController might sometimes think there s new text to speak
after a key event even if there isn't, so we perform an extra check for that.

This prevents Shogun from repeating the credits text on every key press
when navigating the initial menu with menu detection off.

It is kind of an invasive change, so fingers crossed that it won't break anything.