Game and metadata import rework
I made a lot of changes because I thought it was using huge amounts
of memory, but it turns out it was the address sanitizer and Instruments Leaks
diagnostics program that were using all that memory.

- Concurrent imports, as in starting an import or download while another is running,
should be rethought.
- Writes to singleton properties need to be made thread safe.
- The metadata download operations should probably not directly add image download
operations to the queue, but to an array, so that we get more control over how to add a
"finisher" operation at the end (or do we?).