SCOTT: seasofblood.c: Patch DROP HELMET bug
If you drop the helmet in Seas of Blood, it disappears.
There is no way to get it back. However, if you wear
the helmet before dropping it, it ends up in the room
as expected. This seems likely to be a bug, so we patch it,
changing the "destroy" command to "move to current room".